requirements.txt
pillow==9.3.0
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
rembg==2.0.41
numpy==1.23.5